move roleset and parser to separate packages
