+++ lib/services/role.go
@@ -20,27 +20,26 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"sort"
 	"strings"
-	"time"
 
 	"golang.org/x/crypto/ssh"
 
+	"github.com/gravitational/configure/cstrings"
 	"github.com/gravitational/teleport"
 	"github.com/gravitational/teleport/api/types"
 	"github.com/gravitational/teleport/api/types/wrappers"
 	"github.com/gravitational/teleport/lib/defaults"
 	"github.com/gravitational/teleport/lib/modules"
+	"github.com/gravitational/teleport/lib/services/roleset"
+	"github.com/gravitational/teleport/lib/services/ruleset"
 	"github.com/gravitational/teleport/lib/tlsca"
 	"github.com/gravitational/teleport/lib/utils"
 	"github.com/gravitational/teleport/lib/utils/parse"
 
-	"github.com/gravitational/configure/cstrings"
 	"github.com/gravitational/trace"
 
 	"github.com/pborman/uuid"
 	log "github.com/sirupsen/logrus"
-	"github.com/vulcand/predicate"
 )
 
 // getExtendedAdminUserRules provides access to the default set of rules assigned to
@@ -140,32 +139,6 @@ func NewAdminRole() Role {
 	role.SetKubeUsers(Allow, []string{teleport.TraitInternalKubeUsersVariable})
 	role.SetKubeGroups(Allow, []string{teleport.TraitInternalKubeGroupsVariable})
 	return role
-}
-
-// NewImplicitRole is the default implicit role that gets added to all
-// RoleSets.
-func NewImplicitRole() Role {
-	return &RoleV3{
-		Kind:    KindRole,
-		Version: V3,
-		Metadata: Metadata{
-			Name:      teleport.DefaultImplicitRole,
-			Namespace: defaults.Namespace,
-		},
-		Spec: RoleSpecV3{
-			Options: RoleOptions{
-				MaxSessionTTL: MaxDuration(),
-				// PortForwarding has to be set to false in the default-implicit-role
-				// otherwise all roles will be allowed to forward ports (since we default
-				// to true in the check).
-				PortForwarding: NewBoolOption(false),
-			},
-			Allow: RoleConditions{
-				Namespaces: []string{defaults.Namespace},
-				Rules:      CopyRulesSlice(DefaultImplicitRules),
-			},
-		},
-	}
 }
 
 // RoleForUser creates an admin role for a services.User.
@@ -365,7 +338,7 @@ func ValidateRole(r Role) error {
 // validateRule parses the where and action fields to validate the rule.
 func validateRule(r Rule) error {
 	if len(r.Where) != 0 {
-		parser, err := NewWhereParser(&Context{})
+		parser, err := ruleset.NewWhereParser(&Context{})
 		if err != nil {
 			return trace.Wrap(err)
 		}
@@ -376,7 +349,7 @@ func validateRule(r Rule) error {
 	}
 
 	if len(r.Actions) != 0 {
-		parser, err := NewActionsParser(&Context{})
+		parser, err := ruleset.NewActionsParser(&Context{})
 		if err != nil {
 			return trace.Wrap(err)
 		}
@@ -388,6 +361,106 @@ func validateRule(r Rule) error {
 		}
 	}
 	return nil
+}
+
+// FromSpec returns new RoleSet created from spec
+func FromSpec(name string, spec RoleSpecV3) (RoleSet, error) {
+	role, err := NewRole(name, spec)
+	if err != nil {
+		return nil, trace.Wrap(err)
+	}
+
+	return roleset.NewRoleSet(role), nil
+}
+
+// RW is a shortcut that returns all verbs.
+func RW() []string {
+	return []string{VerbList, VerbCreate, VerbRead, VerbUpdate, VerbDelete}
+}
+
+// RO is a shortcut that returns read only verbs that provide access to secrets.
+func RO() []string {
+	return []string{VerbList, VerbRead}
+}
+
+// ReadNoSecrets is a shortcut that returns read only verbs that do not
+// provide access to secrets.
+func ReadNoSecrets() []string {
+	return []string{VerbList, VerbReadNoSecrets}
+}
+
+// RoleGetter is an interface that defines GetRole method
+type RoleGetter interface {
+	// GetRole returns role by name
+	GetRole(ctx context.Context, name string) (Role, error)
+}
+
+// ExtractFromCertificate will extract roles and traits from a *ssh.Certificate
+// or from the backend if they do not exist in the certificate.
+func ExtractFromCertificate(access UserGetter, cert *ssh.Certificate) ([]string, wrappers.Traits, error) {
+	// For legacy certificates, fetch roles and traits from the services.User
+	// object in the backend.
+	if isFormatOld(cert) {
+		u, err := access.GetUser(cert.KeyId, false)
+		if err != nil {
+			return nil, nil, trace.Wrap(err)
+		}
+		log.Warnf("User %v using old style SSH certificate, fetching roles and traits "+
+			"from backend. If the identity provider allows username changes, this can "+
+			"potentially allow an attacker to change the role of the existing user. "+
+			"It's recommended to upgrade to standard SSH certificates.", cert.KeyId)
+		return u.GetRoles(), u.GetTraits(), nil
+	}
+
+	// Standard certificates have the roles and traits embedded in them.
+	roles, err := ExtractRolesFromCert(cert)
+	if err != nil {
+		return nil, nil, trace.Wrap(err)
+	}
+	traits, err := ExtractTraitsFromCert(cert)
+	if err != nil {
+		return nil, nil, trace.Wrap(err)
+	}
+
+	return roles, traits, nil
+}
+
+// ExtractFromIdentity will extract roles and traits from the *x509.Certificate
+// which Teleport passes along as a *tlsca.Identity. If roles and traits do not
+// exist in the certificates, they are extracted from the backend.
+func ExtractFromIdentity(access UserGetter, identity tlsca.Identity) ([]string, wrappers.Traits, error) {
+	// For legacy certificates, fetch roles and traits from the services.User
+	// object in the backend.
+	if missingIdentity(identity) {
+		u, err := access.GetUser(identity.Username, false)
+		if err != nil {
+			return nil, nil, trace.Wrap(err)
+		}
+
+		log.Warnf("Failed to find roles or traits in x509 identity for %v. Fetching	"+
+			"from backend. If the identity provider allows username changes, this can "+
+			"potentially allow an attacker to change the role of the existing user.",
+			identity.Username)
+		return u.GetRoles(), u.GetTraits(), nil
+	}
+
+	return identity.Groups, identity.Traits, nil
+}
+
+// FetchRoleList fetches roles by their names, applies the traits to role
+// variables, and returns the list
+func FetchRoleList(roleNames []string, access RoleGetter, traits map[string][]string) (RoleSet, error) {
+	var roles []Role
+
+	for _, roleName := range roleNames {
+		role, err := access.GetRole(context.TODO(), roleName)
+		if err != nil {
+			return nil, trace.Wrap(err)
+		}
+		roles = append(roles, ApplyTraits(role, traits))
+	}
+
+	return roles, nil
 }
 
 // ApplyTraits applies the passed in traits to any variables within the role
@@ -614,358 +687,15 @@ func ApplyValueTraits(val string, traits map[string][]string) ([]string, error) 
 	return interpolated, nil
 }
 
-// ruleScore is a sorting score of the rule, the larger the score, the more
-// specific the rule is
-func ruleScore(r *Rule) int {
-	score := 0
-	// wildcard rules are less specific
-	if utils.SliceContainsStr(r.Resources, Wildcard) {
-		score -= 4
-	} else if len(r.Resources) == 1 {
-		// rules that match specific resource are more specific than
-		// fields that match several resources
-		score += 2
-	}
-	// rules that have wildcard verbs are less specific
-	if utils.SliceContainsStr(r.Verbs, Wildcard) {
-		score -= 2
-	}
-	// rules that supply 'where' or 'actions' are more specific
-	// having 'where' or 'actions' is more important than
-	// whether the rules are wildcard or not, so here we have +8 vs
-	// -4 and -2 score penalty for wildcards in resources and verbs
-	if len(r.Where) > 0 {
-		score += 8
-	}
-	// rules featuring actions are more specific
-	if len(r.Actions) > 0 {
-		score += 8
-	}
-	return score
-}
-
-// CompareRuleScore returns true if the first rule is more specific than the other.
-//
-// * nRule matching wildcard resource is less specific
-// than same rule matching specific resource.
-// * Rule that has wildcard verbs is less specific
-// than the same rules matching specific verb.
-// * Rule that has where section is more specific
-// than the same rule without where section.
-// * Rule that has actions list is more specific than
-// rule without actions list.
-func CompareRuleScore(r *Rule, o *Rule) bool {
-	return ruleScore(r) > ruleScore(o)
-}
-
-// RuleSet maps resource to a set of rules defined for it
-type RuleSet map[string][]Rule
-
-// MakeRuleSet creates a new rule set from a list
-func MakeRuleSet(rules []Rule) RuleSet {
-	set := make(RuleSet)
-	for _, rule := range rules {
-		for _, resource := range rule.Resources {
-			set[resource] = append(set[resource], rule)
-		}
-	}
-	for resource := range set {
-		rules := set[resource]
-		// sort rules by most specific rule, the rule that has actions
-		// is more specific than the one that has no actions
-		sort.Slice(rules, func(i, j int) bool {
-			return CompareRuleScore(&rules[i], &rules[j])
-		})
-		set[resource] = rules
-	}
-	return set
-}
-
-// Match tests if the resource name and verb are in a given list of rules.
-// More specific rules will be matched first. See Rule.IsMoreSpecificThan
-// for exact specs on whether the rule is more or less specific.
-//
-// Specifying order solves the problem on having multiple rules, e.g. one wildcard
-// rule can override more specific rules with 'where' sections that can have
-// 'actions' lists with side effects that will not be triggered otherwise.
-//
-func (set RuleSet) Match(whereParser predicate.Parser, actionsParser predicate.Parser, resource string, verb string) (bool, error) {
-	// empty set matches nothing
-	if len(set) == 0 {
-		return false, nil
-	}
-
-	// check for matching resource by name
-	// the most specific rule should win
-	rules := set[resource]
-	for _, rule := range rules {
-		match, err := matchesWhere(&rule, whereParser)
-		if err != nil {
-			return false, trace.Wrap(err)
-		}
-		if match && (rule.HasVerb(Wildcard) || rule.HasVerb(verb)) {
-			if err := processActions(&rule, actionsParser); err != nil {
-				return true, trace.Wrap(err)
-			}
-			return true, nil
-		}
-	}
-
-	// check for wildcard resource matcher
-	for _, rule := range set[Wildcard] {
-		match, err := matchesWhere(&rule, whereParser)
-		if err != nil {
-			return false, trace.Wrap(err)
-		}
-		if match && (rule.HasVerb(Wildcard) || rule.HasVerb(verb)) {
-			if err := processActions(&rule, actionsParser); err != nil {
-				return true, trace.Wrap(err)
-			}
-			return true, nil
-		}
-	}
-
-	return false, nil
-}
-
-// matchesWhere returns true if Where rule matches.
-// Empty Where block always matches.
-func matchesWhere(r *Rule, parser predicate.Parser) (bool, error) {
-	if r.Where == "" {
-		return true, nil
-	}
-	ifn, err := parser.Parse(r.Where)
-	if err != nil {
-		return false, trace.Wrap(err)
-	}
-	fn, ok := ifn.(predicate.BoolPredicate)
-	if !ok {
-		return false, trace.BadParameter("invalid predicate type for where expression: %v", r.Where)
-	}
-	return fn(), nil
-}
-
-// processActions processes actions specified for this rule
-func processActions(r *Rule, parser predicate.Parser) error {
-	for _, action := range r.Actions {
-		ifn, err := parser.Parse(action)
-		if err != nil {
-			return trace.Wrap(err)
-		}
-		fn, ok := ifn.(predicate.BoolPredicate)
-		if !ok {
-			return trace.BadParameter("invalid predicate type for action expression: %v", action)
-		}
-		fn()
-	}
-	return nil
-}
-
-// Slice returns slice from a set
-func (set RuleSet) Slice() []Rule {
-	var out []Rule
-	for _, rules := range set {
-		out = append(out, rules...)
-	}
-	return out
-}
-
-// AccessChecker interface implements access checks for given role or role set
-type AccessChecker interface {
-	// HasRole checks if the checker includes the role
-	HasRole(role string) bool
-
-	// RoleNames returns a list of role names
-	RoleNames() []string
-
-	// CheckAccessToServer checks access to server.
-	CheckAccessToServer(login string, server Server, mfa AccessMFAParams) error
-
-	// CheckAccessToRemoteCluster checks access to remote cluster
-	CheckAccessToRemoteCluster(cluster RemoteCluster) error
-
-	// CheckAccessToRule checks access to a rule within a namespace.
-	CheckAccessToRule(context RuleContext, namespace string, rule string, verb string, silent bool) error
-
-	// CheckLoginDuration checks if role set can login up to given duration and
-	// returns a combined list of allowed logins.
-	CheckLoginDuration(ttl time.Duration) ([]string, error)
-
-	// CheckKubeGroupsAndUsers check if role can login into kubernetes
-	// and returns two lists of combined allowed groups and users
-	CheckKubeGroupsAndUsers(ttl time.Duration, overrideTTL bool) (groups []string, users []string, err error)
-
-	// AdjustSessionTTL will reduce the requested ttl to lowest max allowed TTL
-	// for this role set, otherwise it returns ttl unchanged
-	AdjustSessionTTL(ttl time.Duration) time.Duration
-
-	// AdjustClientIdleTimeout adjusts requested idle timeout
-	// to the lowest max allowed timeout, the most restrictive
-	// option will be picked
-	AdjustClientIdleTimeout(ttl time.Duration) time.Duration
-
-	// AdjustDisconnectExpiredCert adjusts the value based on the role set
-	// the most restrictive option will be picked
-	AdjustDisconnectExpiredCert(disconnect bool) bool
-
-	// CheckAgentForward checks if the role can request agent forward for this
-	// user.
-	CheckAgentForward(login string) error
-
-	// CanForwardAgents returns true if this role set offers capability to forward
-	// agents.
-	CanForwardAgents() bool
-
-	// CanPortForward returns true if this RoleSet can forward ports.
-	CanPortForward() bool
-
-	// MaybeCanReviewRequests attempts to guess if this RoleSet belongs
-	// to a user who should be submitting access reviews. Because not all rolesets
-	// are derived from statically assigned roles, this may return false positives.
-	MaybeCanReviewRequests() bool
-
-	// PermitX11Forwarding returns true if this RoleSet allows X11 Forwarding.
-	PermitX11Forwarding() bool
-
-	// CertificateFormat returns the most permissive certificate format in a
-	// RoleSet.
-	CertificateFormat() string
-
-	// EnhancedRecordingSet returns a set of events that will be recorded
-	// for enhanced session recording.
-	EnhancedRecordingSet() map[string]bool
-
-	// CheckAccessToApp checks access to an application.
-	CheckAccessToApp(login string, app *App, mfa AccessMFAParams) error
-
-	// CheckAccessToKubernetes checks access to a kubernetes cluster.
-	CheckAccessToKubernetes(login string, app *KubernetesCluster, mfa AccessMFAParams) error
-
-	// CheckDatabaseNamesAndUsers returns database names and users this role
-	// is allowed to use.
-	CheckDatabaseNamesAndUsers(ttl time.Duration, overrideTTL bool) (names []string, users []string, err error)
-
-	// CheckAccessToDatabase checks whether a user has access to the provided
-	// database server.
-	CheckAccessToDatabase(server types.DatabaseServer, mfa AccessMFAParams, matchers ...RoleMatcher) error
-
-	// CheckImpersonate checks whether current user is allowed to impersonate
-	// users and roles
-	CheckImpersonate(currentUser, impersonateUser types.User, impersonateRoles []types.Role) error
-
-	// CanImpersonateSomeone returns true if this checker has any impersonation rules
-	CanImpersonateSomeone() bool
-}
-
-// FromSpec returns new RoleSet created from spec
-func FromSpec(name string, spec RoleSpecV3) (RoleSet, error) {
-	role, err := NewRole(name, spec)
-	if err != nil {
-		return nil, trace.Wrap(err)
-	}
-
-	return NewRoleSet(role), nil
-}
-
-// RW is a shortcut that returns all verbs.
-func RW() []string {
-	return []string{VerbList, VerbCreate, VerbRead, VerbUpdate, VerbDelete}
-}
-
-// RO is a shortcut that returns read only verbs that provide access to secrets.
-func RO() []string {
-	return []string{VerbList, VerbRead}
-}
-
-// ReadNoSecrets is a shortcut that returns read only verbs that do not
-// provide access to secrets.
-func ReadNoSecrets() []string {
-	return []string{VerbList, VerbReadNoSecrets}
-}
-
-// RoleGetter is an interface that defines GetRole method
-type RoleGetter interface {
-	// GetRole returns role by name
-	GetRole(ctx context.Context, name string) (Role, error)
-}
-
-// ExtractFromCertificate will extract roles and traits from a *ssh.Certificate
-// or from the backend if they do not exist in the certificate.
-func ExtractFromCertificate(access UserGetter, cert *ssh.Certificate) ([]string, wrappers.Traits, error) {
-	// For legacy certificates, fetch roles and traits from the services.User
-	// object in the backend.
-	if isFormatOld(cert) {
-		u, err := access.GetUser(cert.KeyId, false)
-		if err != nil {
-			return nil, nil, trace.Wrap(err)
-		}
-		log.Warnf("User %v using old style SSH certificate, fetching roles and traits "+
-			"from backend. If the identity provider allows username changes, this can "+
-			"potentially allow an attacker to change the role of the existing user. "+
-			"It's recommended to upgrade to standard SSH certificates.", cert.KeyId)
-		return u.GetRoles(), u.GetTraits(), nil
-	}
-
-	// Standard certificates have the roles and traits embedded in them.
-	roles, err := ExtractRolesFromCert(cert)
-	if err != nil {
-		return nil, nil, trace.Wrap(err)
-	}
-	traits, err := ExtractTraitsFromCert(cert)
-	if err != nil {
-		return nil, nil, trace.Wrap(err)
-	}
-
-	return roles, traits, nil
-}
-
-// ExtractFromIdentity will extract roles and traits from the *x509.Certificate
-// which Teleport passes along as a *tlsca.Identity. If roles and traits do not
-// exist in the certificates, they are extracted from the backend.
-func ExtractFromIdentity(access UserGetter, identity tlsca.Identity) ([]string, wrappers.Traits, error) {
-	// For legacy certificates, fetch roles and traits from the services.User
-	// object in the backend.
-	if missingIdentity(identity) {
-		u, err := access.GetUser(identity.Username, false)
-		if err != nil {
-			return nil, nil, trace.Wrap(err)
-		}
-
-		log.Warnf("Failed to find roles or traits in x509 identity for %v. Fetching	"+
-			"from backend. If the identity provider allows username changes, this can "+
-			"potentially allow an attacker to change the role of the existing user.",
-			identity.Username)
-		return u.GetRoles(), u.GetTraits(), nil
-	}
-
-	return identity.Groups, identity.Traits, nil
-}
-
-// FetchRoleList fetches roles by their names, applies the traits to role
-// variables, and returns the list
-func FetchRoleList(roleNames []string, access RoleGetter, traits map[string][]string) (RoleSet, error) {
-	var roles []Role
-
-	for _, roleName := range roleNames {
-		role, err := access.GetRole(context.TODO(), roleName)
-		if err != nil {
-			return nil, trace.Wrap(err)
-		}
-		roles = append(roles, ApplyTraits(role, traits))
-	}
-
-	return roles, nil
-}
-
 // FetchRoles fetches roles by their names, applies the traits to role
 // variables, and returns the RoleSet. Adds runtime roles like the default
 // implicit role to RoleSet.
 func FetchRoles(roleNames []string, access RoleGetter, traits map[string][]string) (RoleSet, error) {
-	roles, err := FetchRoleList(roleNames, access, traits)
+	roleList, err := FetchRoleList(roleNames, access, traits)
 	if err != nil {
 		return nil, trace.Wrap(err)
 	}
-	return NewRoleSet(roles...), nil
+	return roleset.NewRoleSet(roleList...), nil
 }
 
 // isFormatOld returns true if roles and traits were not found in the
@@ -1010,1110 +740,6 @@ func ExtractTraitsFromCert(cert *ssh.Certificate) (wrappers.Traits, error) {
 		return nil, trace.Wrap(err)
 	}
 	return traits, nil
-}
-
-// NewRoleSet returns new RoleSet based on the roles
-func NewRoleSet(roles ...Role) RoleSet {
-	// unauthenticated Nop role should not have any privileges
-	// by default, otherwise it is too permissive
-	if len(roles) == 1 && roles[0].GetName() == string(teleport.RoleNop) {
-		return roles
-	}
-	return append(roles, NewImplicitRole())
-}
-
-// RoleSet is a set of roles that implements access control functionality
-type RoleSet []Role
-
-// MatchNamespace returns true if given list of namespace matches
-// target namespace, wildcard matches everything.
-func MatchNamespace(selectors []string, namespace string) (bool, string) {
-	for _, n := range selectors {
-		if n == namespace || n == Wildcard {
-			return true, "matched"
-		}
-	}
-	return false, fmt.Sprintf("no match, role selectors %v, server namespace: %v", selectors, namespace)
-}
-
-// MatchLogin returns true if attempted login matches any of the logins.
-func MatchLogin(selectors []string, login string) (bool, string) {
-	for _, l := range selectors {
-		if l == login {
-			return true, "matched"
-		}
-	}
-	return false, fmt.Sprintf("no match, role selectors %v, login: %v", selectors, login)
-}
-
-// MatchDatabaseName returns true if provided database name matches selectors.
-func MatchDatabaseName(selectors []string, name string) (bool, string) {
-	for _, n := range selectors {
-		if n == name || n == Wildcard {
-			return true, "matched"
-		}
-	}
-	return false, fmt.Sprintf("no match, role selectors %v, database name: %v", selectors, name)
-}
-
-// MatchDatabaseUser returns true if provided database user matches selectors.
-func MatchDatabaseUser(selectors []string, user string) (bool, string) {
-	for _, u := range selectors {
-		if u == user || u == Wildcard {
-			return true, "matched"
-		}
-	}
-	return false, fmt.Sprintf("no match, role selectors %v, database user: %v", selectors, user)
-}
-
-// MatchLabels matches selector against target. Empty selector matches
-// nothing, wildcard matches everything.
-func MatchLabels(selector Labels, target map[string]string) (bool, string, error) {
-	// Empty selector matches nothing.
-	if len(selector) == 0 {
-		return false, "no match, empty selector", nil
-	}
-
-	// *: * matches everything even empty target set.
-	selectorValues := selector[Wildcard]
-	if len(selectorValues) == 1 && selectorValues[0] == Wildcard {
-		return true, "matched", nil
-	}
-
-	// Perform full match.
-	for key, selectorValues := range selector {
-		targetVal, hasKey := target[key]
-
-		if !hasKey {
-			return false, fmt.Sprintf("no key match: '%v'", key), nil
-		}
-
-		if !utils.SliceContainsStr(selectorValues, Wildcard) {
-			result, err := utils.SliceMatchesRegex(targetVal, selectorValues)
-			if err != nil {
-				return false, "", trace.Wrap(err)
-			} else if !result {
-				return false, fmt.Sprintf("no value match: got '%v' want: '%v'", targetVal, selectorValues), nil
-			}
-		}
-	}
-
-	return true, "matched", nil
-}
-
-// RoleNames returns a slice with role names. Removes runtime roles like
-// the default implicit role.
-func (set RoleSet) RoleNames() []string {
-	out := make([]string, 0, len(set))
-	for _, r := range set {
-		if r.GetName() == teleport.DefaultImplicitRole {
-			continue
-		}
-		out = append(out, r.GetName())
-	}
-	return out
-}
-
-// HasRole checks if the role set has the role
-func (set RoleSet) HasRole(role string) bool {
-	for _, r := range set {
-		if r.GetName() == role {
-			return true
-		}
-	}
-	return false
-}
-
-// AdjustSessionTTL will reduce the requested ttl to lowest max allowed TTL
-// for this role set, otherwise it returns ttl unchanged
-func (set RoleSet) AdjustSessionTTL(ttl time.Duration) time.Duration {
-	for _, role := range set {
-		maxSessionTTL := role.GetOptions().MaxSessionTTL.Value()
-		if maxSessionTTL != 0 && ttl > maxSessionTTL {
-			ttl = maxSessionTTL
-		}
-	}
-	return ttl
-}
-
-// MaxConnections returns the maximum number of concurrent ssh connections
-// allowed.  If MaxConnections is zero then no maximum was defined
-// and the number of concurrent connections is unconstrained.
-func (set RoleSet) MaxConnections() int64 {
-	var mcs int64
-	for _, role := range set {
-		if m := role.GetOptions().MaxConnections; m != 0 && (m < mcs || mcs == 0) {
-			mcs = m
-		}
-	}
-	return mcs
-}
-
-// MaxSessions returns the maximum number of concurrent ssh sessions
-// per connection.  If MaxSessions is zero then no maximum was defined
-// and the number of sessions is unconstrained.
-func (set RoleSet) MaxSessions() int64 {
-	var ms int64
-	for _, role := range set {
-		if m := role.GetOptions().MaxSessions; m != 0 && (m < ms || ms == 0) {
-			ms = m
-		}
-	}
-	return ms
-}
-
-// AdjustClientIdleTimeout adjusts requested idle timeout
-// to the lowest max allowed timeout, the most restrictive
-// option will be picked, negative values will be assumed as 0
-func (set RoleSet) AdjustClientIdleTimeout(timeout time.Duration) time.Duration {
-	if timeout < 0 {
-		timeout = 0
-	}
-	for _, role := range set {
-		roleTimeout := role.GetOptions().ClientIdleTimeout
-		// 0 means not set, so it can't be most restrictive, disregard it too
-		if roleTimeout.Duration() <= 0 {
-			continue
-		}
-		switch {
-		// in case if timeout is 0, means that incoming value
-		// does not restrict the idle timeout, pick any other value
-		// set by the role
-		case timeout == 0:
-			timeout = roleTimeout.Duration()
-		case roleTimeout.Duration() < timeout:
-			timeout = roleTimeout.Duration()
-		}
-	}
-	return timeout
-}
-
-// AdjustDisconnectExpiredCert adjusts the value based on the role set
-// the most restrictive option will be picked
-func (set RoleSet) AdjustDisconnectExpiredCert(disconnect bool) bool {
-	for _, role := range set {
-		if role.GetOptions().DisconnectExpiredCert.Value() {
-			disconnect = true
-		}
-	}
-	return disconnect
-}
-
-// CheckKubeGroupsAndUsers check if role can login into kubernetes
-// and returns two lists of allowed groups and users
-func (set RoleSet) CheckKubeGroupsAndUsers(ttl time.Duration, overrideTTL bool) ([]string, []string, error) {
-	groups := make(map[string]struct{})
-	users := make(map[string]struct{})
-	var matchedTTL bool
-	for _, role := range set {
-		maxSessionTTL := role.GetOptions().MaxSessionTTL.Value()
-		if overrideTTL || (ttl <= maxSessionTTL && maxSessionTTL != 0) {
-			matchedTTL = true
-			for _, group := range role.GetKubeGroups(Allow) {
-				groups[group] = struct{}{}
-			}
-			for _, user := range role.GetKubeUsers(Allow) {
-				users[user] = struct{}{}
-			}
-		}
-	}
-	for _, role := range set {
-		for _, group := range role.GetKubeGroups(Deny) {
-			delete(groups, group)
-		}
-		for _, user := range role.GetKubeUsers(Deny) {
-			delete(users, user)
-		}
-	}
-	if !matchedTTL {
-		return nil, nil, trace.AccessDenied("this user cannot request kubernetes access for %v", ttl)
-	}
-	if len(groups) == 0 && len(users) == 0 {
-		return nil, nil, trace.NotFound("this user cannot request kubernetes access, has no assigned groups or users")
-	}
-	return utils.StringsSliceFromSet(groups), utils.StringsSliceFromSet(users), nil
-}
-
-// CheckDatabaseNamesAndUsers checks if the role has any allowed database
-// names or users.
-func (set RoleSet) CheckDatabaseNamesAndUsers(ttl time.Duration, overrideTTL bool) ([]string, []string, error) {
-	names := make(map[string]struct{})
-	users := make(map[string]struct{})
-	var matchedTTL bool
-	for _, role := range set {
-		maxSessionTTL := role.GetOptions().MaxSessionTTL.Value()
-		if overrideTTL || (ttl <= maxSessionTTL && maxSessionTTL != 0) {
-			matchedTTL = true
-			for _, name := range role.GetDatabaseNames(Allow) {
-				names[name] = struct{}{}
-			}
-			for _, user := range role.GetDatabaseUsers(Allow) {
-				users[user] = struct{}{}
-			}
-		}
-	}
-	for _, role := range set {
-		for _, name := range role.GetDatabaseNames(Deny) {
-			delete(names, name)
-		}
-		for _, user := range role.GetDatabaseUsers(Deny) {
-			delete(users, user)
-		}
-	}
-	if !matchedTTL {
-		return nil, nil, trace.AccessDenied("this user cannot request database access for %v", ttl)
-	}
-	if len(names) == 0 && len(users) == 0 {
-		return nil, nil, trace.NotFound("this user cannot request database access, has no assigned database names or users")
-	}
-	return utils.StringsSliceFromSet(names), utils.StringsSliceFromSet(users), nil
-}
-
-// CheckLoginDuration checks if role set can login up to given duration and
-// returns a combined list of allowed logins.
-func (set RoleSet) CheckLoginDuration(ttl time.Duration) ([]string, error) {
-	logins, matchedTTL := set.GetLoginsForTTL(ttl)
-	if !matchedTTL {
-		return nil, trace.AccessDenied("this user cannot request a certificate for %v", ttl)
-	}
-
-	if len(logins) == 0 && !set.hasPossibleLogins() {
-		// user was deliberately configured to have no login capability,
-		// but ssh certificates must contain at least one valid principal.
-		// we add a single distinctive value which should be unique, and
-		// will never be a valid unix login (due to leading '-').
-		logins = []string{"-teleport-nologin-" + uuid.New()}
-	}
-
-	if len(logins) == 0 {
-		return nil, trace.AccessDenied("this user cannot create SSH sessions, has no allowed logins")
-	}
-
-	return logins, nil
-}
-
-// GetLoginsForTTL collects all logins that are valid for the given TTL.  The matchedTTL
-// value indicates whether the TTL is within scope of *any* role.  This helps to distinguish
-// between TTLs which are categorically invalid, and TTLs which are theoretically valid
-// but happen to grant no logins.
-func (set RoleSet) GetLoginsForTTL(ttl time.Duration) (logins []string, matchedTTL bool) {
-	for _, role := range set {
-		maxSessionTTL := role.GetOptions().MaxSessionTTL.Value()
-		if ttl <= maxSessionTTL && maxSessionTTL != 0 {
-			matchedTTL = true
-			logins = append(logins, role.GetLogins(Allow)...)
-		}
-	}
-	return utils.Deduplicate(logins), matchedTTL
-}
-
-// CheckAccessToRemoteCluster checks if a role has access to remote cluster. Deny rules are
-// checked first then allow rules. Access to a cluster is determined by
-// namespaces, labels, and logins.
-//
-// Note, logging in this function only happens in debug mode, this is because
-// adding logging to this function (which is called on every server returned
-// by GetRemoteClusters) can slow down this function by 50x for large clusters!
-func (set RoleSet) CheckAccessToRemoteCluster(rc RemoteCluster) error {
-	if len(set) == 0 {
-		return trace.AccessDenied("access to cluster denied")
-	}
-
-	var errs []error
-
-	rcLabels := rc.GetMetadata().Labels
-
-	// For backwards compatibility, if there is no role in the set with labels and the cluster
-	// has no labels, assume that the role set has access to the cluster.
-	usesLabels := false
-	for _, role := range set {
-		if len(role.GetClusterLabels(Allow)) != 0 || len(role.GetClusterLabels(Deny)) != 0 {
-			usesLabels = true
-			break
-		}
-	}
-
-	if usesLabels == false && len(rcLabels) == 0 {
-		if log.GetLevel() == log.DebugLevel {
-			log.WithFields(log.Fields{
-				trace.Component: teleport.ComponentRBAC,
-			}).Debugf("Grant access to cluster %v - no role in %v uses cluster labels and the cluster is not labeled.",
-				rc.GetName(), set.RoleNames())
-		}
-		return nil
-	}
-
-	// Check deny rules first: a single matching label from
-	// the deny role set prohibits access.
-	for _, role := range set {
-		matchLabels, labelsMessage, err := MatchLabels(role.GetClusterLabels(Deny), rcLabels)
-		if err != nil {
-			return trace.Wrap(err)
-		}
-		if matchLabels {
-			// This condition avoids formatting calls on large scale.
-			if log.GetLevel() == log.DebugLevel {
-				log.WithFields(log.Fields{
-					trace.Component: teleport.ComponentRBAC,
-				}).Debugf("Access to cluster %v denied, deny rule in %v matched; match(label=%v)",
-					rc.GetName(), role.GetName(), labelsMessage)
-			}
-			return trace.AccessDenied("access to cluster denied")
-		}
-	}
-
-	// Check allow rules: label has to match in any role in the role set to be granted access.
-	for _, role := range set {
-		matchLabels, labelsMessage, err := MatchLabels(role.GetClusterLabels(Allow), rcLabels)
-		log.WithFields(log.Fields{
-			trace.Component: teleport.ComponentRBAC,
-		}).Debugf("Check access to role(%v) rc(%v, labels=%v) matchLabels=%v, msg=%v, err=%v allow=%v rcLabels=%v",
-			role.GetName(), rc.GetName(), rcLabels, matchLabels, labelsMessage, err, role.GetClusterLabels(Allow), rcLabels)
-		if err != nil {
-			return trace.Wrap(err)
-		}
-		if matchLabels {
-			return nil
-		}
-		if log.GetLevel() == log.DebugLevel {
-			deniedError := trace.AccessDenied("role=%v, match(label=%v)",
-				role.GetName(), labelsMessage)
-			errs = append(errs, deniedError)
-		}
-	}
-
-	if log.GetLevel() == log.DebugLevel {
-		log.WithFields(log.Fields{
-			trace.Component: teleport.ComponentRBAC,
-		}).Debugf("Access to cluster %v denied, no allow rule matched; %v", rc.GetName(), errs)
-	}
-	return trace.AccessDenied("access to cluster denied")
-}
-
-func (set RoleSet) hasPossibleLogins() bool {
-	for _, role := range set {
-		if role.GetName() == teleport.DefaultImplicitRole {
-			continue
-		}
-		if len(role.GetLogins(Allow)) != 0 {
-			return true
-		}
-	}
-	return false
-}
-
-// CheckAccessToServer checks if a role has access to a node. Deny rules are
-// checked first then allow rules. Access to a node is determined by
-// namespaces, labels, and logins.
-//
-// Note, logging in this function only happens in debug mode, this is because
-// adding logging to this function (which is called on every server returned
-// by GetNodes) can slow down this function by 50x for large clusters!
-func (set RoleSet) CheckAccessToServer(login string, s Server, mfa AccessMFAParams) error {
-	if mfa.AlwaysRequired && !mfa.Verified {
-		log.WithFields(log.Fields{
-			trace.Component: teleport.ComponentRBAC,
-		}).Debugf("Access to node %q denied, cluster requires per-session MFA", s.GetHostname())
-		return ErrSessionMFARequired
-	}
-	var errs []error
-
-	// Check deny rules first: a single matching namespace, label, or login from
-	// the deny role set prohibits access.
-	for _, role := range set {
-		matchNamespace, namespaceMessage := MatchNamespace(role.GetNamespaces(Deny), s.GetNamespace())
-		matchLabels, labelsMessage, err := MatchLabels(role.GetNodeLabels(Deny), s.GetAllLabels())
-		if err != nil {
-			return trace.Wrap(err)
-		}
-		matchLogin, loginMessage := MatchLogin(role.GetLogins(Deny), login)
-		if matchNamespace && (matchLabels || matchLogin) {
-			if log.GetLevel() == log.DebugLevel {
-				log.WithFields(log.Fields{
-					trace.Component: teleport.ComponentRBAC,
-				}).Debugf("Access to node %v denied, deny rule in %v matched; match(namespace=%v, label=%v, login=%v)",
-					s.GetHostname(), role.GetName(), namespaceMessage, labelsMessage, loginMessage)
-			}
-			return trace.AccessDenied("access to server denied")
-		}
-	}
-
-	allowed := false
-	// Check allow rules: namespace, label, and login have to all match in
-	// one role in the role set to be granted access.
-	for _, role := range set {
-		matchNamespace, namespaceMessage := MatchNamespace(role.GetNamespaces(Allow), s.GetNamespace())
-		matchLabels, labelsMessage, err := MatchLabels(role.GetNodeLabels(Allow), s.GetAllLabels())
-		if err != nil {
-			return trace.Wrap(err)
-		}
-		matchLogin, loginMessage := MatchLogin(role.GetLogins(Allow), login)
-		if matchNamespace && matchLabels && matchLogin {
-			if mfa.Verified {
-				return nil
-			}
-			if role.GetOptions().RequireSessionMFA {
-				log.WithFields(log.Fields{
-					trace.Component: teleport.ComponentRBAC,
-				}).Debugf("Access to node %q denied, role %q requires per-session MFA; match(namespace=%v, label=%v, login=%v)",
-					s.GetHostname(), role.GetName(), namespaceMessage, labelsMessage, loginMessage)
-				return ErrSessionMFARequired
-			}
-			// Check all remaining roles, even if we found a match.
-			// RequireSessionMFA should be enforced when at least one role has
-			// it.
-			allowed = true
-			continue
-		}
-		if log.GetLevel() == log.DebugLevel {
-			deniedError := trace.AccessDenied("role=%v, match(namespace=%v, label=%v, login=%v)",
-				role.GetName(), namespaceMessage, labelsMessage, loginMessage)
-			errs = append(errs, deniedError)
-		}
-	}
-	if allowed {
-		return nil
-	}
-
-	if log.GetLevel() == log.DebugLevel {
-		log.WithFields(log.Fields{
-			trace.Component: teleport.ComponentRBAC,
-		}).Debugf("Access to node %v denied, no allow rule matched; %v", s.GetHostname(), errs)
-	}
-	return trace.AccessDenied("access to server denied")
-}
-
-// CheckAccessToApp checks if a role has access to an application. Deny rules
-// are checked first, then allow rules. Access to an application is determined by
-// namespaces and labels.
-func (set RoleSet) CheckAccessToApp(namespace string, app *App, mfa AccessMFAParams) error {
-	if mfa.AlwaysRequired && !mfa.Verified {
-		log.WithFields(log.Fields{
-			trace.Component: teleport.ComponentRBAC,
-		}).Debugf("Access to app %q denied, cluster requires per-session MFA", app.Name)
-		return ErrSessionMFARequired
-	}
-	var errs []error
-
-	// Check deny rules: a matching namespace and label in the deny section
-	// prohibits access.
-	for _, role := range set {
-		matchNamespace, namespaceMessage := MatchNamespace(role.GetNamespaces(Deny), namespace)
-		matchLabels, labelsMessage, err := MatchLabels(role.GetAppLabels(Deny), CombineLabels(app.StaticLabels, app.DynamicLabels))
-		if err != nil {
-			return trace.Wrap(err)
-		}
-		if matchNamespace && matchLabels {
-			if log.GetLevel() == log.DebugLevel {
-				log.WithFields(log.Fields{
-					trace.Component: teleport.ComponentRBAC,
-				}).Debugf("Access to app %v denied, deny rule in %v matched; match(namespace=%v, label=%v)",
-					app.Name, role.GetName(), namespaceMessage, labelsMessage)
-			}
-			return trace.AccessDenied("access to app denied")
-		}
-	}
-
-	allowed := false
-	// Check allow rules: namespace and label both have to match to be granted access.
-	for _, role := range set {
-		matchNamespace, namespaceMessage := MatchNamespace(role.GetNamespaces(Allow), namespace)
-		matchLabels, labelsMessage, err := MatchLabels(role.GetAppLabels(Allow), CombineLabels(app.StaticLabels, app.DynamicLabels))
-		if err != nil {
-			return trace.Wrap(err)
-		}
-		if matchNamespace && matchLabels {
-			if mfa.Verified {
-				return nil
-			}
-			if role.GetOptions().RequireSessionMFA {
-				log.WithFields(log.Fields{
-					trace.Component: teleport.ComponentRBAC,
-				}).Debugf("Access to app %q denied, role %q requires per-session MFA; match(namespace=%v, label=%v)",
-					app.Name, role.GetName(), namespaceMessage, labelsMessage)
-				return ErrSessionMFARequired
-			}
-			// Check all remaining roles, even if we found a match.
-			// RequireSessionMFA should be enforced when at least one role has
-			// it.
-			allowed = true
-			continue
-		}
-		if log.GetLevel() == log.DebugLevel {
-			deniedError := trace.AccessDenied("role=%v, match(namespace=%v, label=%v)",
-				role.GetName(), namespaceMessage, labelsMessage)
-			errs = append(errs, deniedError)
-		}
-	}
-	if allowed {
-		return nil
-	}
-
-	if log.GetLevel() == log.DebugLevel {
-		log.WithFields(log.Fields{
-			trace.Component: teleport.ComponentRBAC,
-		}).Debugf("Access to app %v denied, no allow rule matched; %v", app.Name, errs)
-	}
-	return trace.AccessDenied("access to app denied")
-}
-
-// CheckAccessToKubernetes checks if a role has access to a kubernetes cluster.
-// Deny rules are checked first, then allow rules. Access to a kubernetes
-// cluster is determined by namespaces and labels.
-func (set RoleSet) CheckAccessToKubernetes(namespace string, kube *KubernetesCluster, mfa AccessMFAParams) error {
-	if mfa.AlwaysRequired && !mfa.Verified {
-		log.WithFields(log.Fields{
-			trace.Component: teleport.ComponentRBAC,
-		}).Debugf("Access to kubernetes cluster %q denied, cluster requires per-session MFA", kube.Name)
-		return ErrSessionMFARequired
-	}
-	var errs []error
-
-	// Check deny rules: a matching namespace and label in the deny section
-	// prohibits access.
-	for _, role := range set {
-		matchNamespace, namespaceMessage := MatchNamespace(role.GetNamespaces(Deny), namespace)
-		matchLabels, labelsMessage, err := MatchLabels(role.GetKubernetesLabels(Deny), CombineLabels(kube.StaticLabels, kube.DynamicLabels))
-		if err != nil {
-			return trace.Wrap(err)
-		}
-		if matchNamespace && matchLabels {
-			if log.GetLevel() == log.DebugLevel {
-				log.WithFields(log.Fields{
-					trace.Component: teleport.ComponentRBAC,
-				}).Debugf("Access to kubernetes cluster %v denied, deny rule in %v matched; match(namespace=%v, label=%v)",
-					kube.Name, role.GetName(), namespaceMessage, labelsMessage)
-			}
-			return trace.AccessDenied("access to kubernetes cluster denied")
-		}
-	}
-
-	allowed := false
-	// Check allow rules: namespace and label both have to match to be granted access.
-	for _, role := range set {
-		matchNamespace, namespaceMessage := MatchNamespace(role.GetNamespaces(Allow), namespace)
-		matchLabels, labelsMessage, err := MatchLabels(role.GetKubernetesLabels(Allow), CombineLabels(kube.StaticLabels, kube.DynamicLabels))
-		if err != nil {
-			return trace.Wrap(err)
-		}
-		if matchNamespace && matchLabels {
-			if mfa.Verified {
-				return nil
-			}
-			if role.GetOptions().RequireSessionMFA {
-				log.WithFields(log.Fields{
-					trace.Component: teleport.ComponentRBAC,
-				}).Debugf("Access to kubernetes cluster %q denied, role %q requires per-session MFA; match(namespace=%v, label=%v)",
-					kube.Name, role.GetName(), namespaceMessage, labelsMessage)
-				return ErrSessionMFARequired
-			}
-			// Check all remaining roles, even if we found a match.
-			// RequireSessionMFA should be enforced when at least one role has
-			// it.
-			allowed = true
-			continue
-		}
-		if log.GetLevel() == log.DebugLevel {
-			deniedError := trace.AccessDenied("role=%v, match(namespace=%v, label=%v)",
-				role.GetName(), namespaceMessage, labelsMessage)
-			errs = append(errs, deniedError)
-		}
-	}
-	if allowed {
-		return nil
-	}
-
-	if log.GetLevel() == log.DebugLevel {
-		log.WithFields(log.Fields{
-			trace.Component: teleport.ComponentRBAC,
-		}).Debugf("Access to kubernetes cluster %v denied, no allow rule matched; %v", kube.Name, errs)
-	}
-	return trace.AccessDenied("access to kubernetes cluster denied")
-}
-
-// CanImpersonateSomeone returns true if this checker has any impersonation rules
-func (set RoleSet) CanImpersonateSomeone() bool {
-	for _, role := range set {
-		cond := role.GetImpersonateConditions(Allow)
-		if !cond.IsEmpty() {
-			return true
-		}
-	}
-	return false
-}
-
-// CheckImpersonate returns nil if this role set can impersonate
-// a user and their roles, returns AccessDenied otherwise
-// CheckImpersonate checks whether current user is allowed to impersonate
-// users and roles
-func (set RoleSet) CheckImpersonate(currentUser, impersonateUser types.User, impersonateRoles []types.Role) error {
-	ctx := &impersonateContext{
-		user:            currentUser,
-		impersonateUser: impersonateUser,
-	}
-	whereParser, err := newImpersonateWhereParser(ctx)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-	// check deny: a single match on a deny rule prohibits access
-	for _, role := range set {
-		cond := role.GetImpersonateConditions(Deny)
-		matched, err := matchDenyImpersonateCondition(cond, impersonateUser, impersonateRoles)
-		if err != nil {
-			return trace.Wrap(err)
-		}
-		if matched {
-			return trace.AccessDenied("access denied to '%s' to impersonate user '%s' and roles '%s'", currentUser.GetName(), impersonateUser.GetName(), roleNames(impersonateRoles))
-		}
-	}
-
-	// check allow: if matches, allow to impersonate
-	for _, role := range set {
-		cond := role.GetImpersonateConditions(Allow)
-		matched, err := matchAllowImpersonateCondition(ctx, whereParser, cond, impersonateUser, impersonateRoles)
-		if err != nil {
-			return trace.Wrap(err)
-		}
-		if matched {
-			return nil
-		}
-	}
-
-	return trace.AccessDenied("access denied to '%s' to impersonate user '%s' and roles '%s'", currentUser.GetName(), impersonateUser.GetName(), roleNames(impersonateRoles))
-}
-
-func roleNames(roles []types.Role) string {
-	out := make([]string, len(roles))
-	for i := range roles {
-		out[i] = roles[i].GetName()
-	}
-	return strings.Join(out, ", ")
-}
-
-// matchAllowImpersonateCondition matches impersonate condition,
-// both user, role and where condition has to match
-func matchAllowImpersonateCondition(ctx *impersonateContext, whereParser predicate.Parser, cond types.ImpersonateConditions, impersonateUser types.User, impersonateRoles []types.Role) (bool, error) {
-	// an empty set matches nothing
-	if len(cond.Users) == 0 && len(cond.Roles) == 0 {
-		return false, nil
-	}
-	// should specify both roles and users, this condition is also verified on the role level
-	if len(cond.Users) == 0 || len(cond.Roles) == 0 {
-		return false, trace.BadParameter("the system does not support empty roles and users")
-	}
-
-	anyUser, err := parse.NewAnyMatcher(cond.Users)
-	if err != nil {
-		return false, trace.Wrap(err)
-	}
-
-	if !anyUser.Match(impersonateUser.GetName()) {
-		return false, nil
-	}
-
-	anyRole, err := parse.NewAnyMatcher(cond.Roles)
-	if err != nil {
-		return false, trace.Wrap(err)
-	}
-
-	for _, impersonateRole := range impersonateRoles {
-		if !anyRole.Match(impersonateRole.GetName()) {
-			return false, nil
-		}
-		// TODO:
-		// This set impersonateRole inside the ctx that is in turn used inside whereParser
-		// which is created in CheckImpersonate above but is being used right below.
-		// This is unfortunate interface of the parser, instead
-		// parser should accept additional context as a first argument.
-		ctx.impersonateRole = impersonateRole
-		match, err := matchesImpersonateWhere(cond, whereParser)
-		if err != nil {
-			return false, trace.Wrap(err)
-		}
-		if !match {
-			return false, nil
-		}
-	}
-
-	return true, nil
-}
-
-// matchDenyImpersonateCondition matches impersonate condition,
-// greedy is used for deny type rules, where any user or role can match
-func matchDenyImpersonateCondition(cond types.ImpersonateConditions, impersonateUser types.User, impersonateRoles []types.Role) (bool, error) {
-	// an empty set matches nothing
-	if len(cond.Users) == 0 && len(cond.Roles) == 0 {
-		return false, nil
-	}
-	// should specify both roles and users, this condition is also verified on the role level
-	if len(cond.Users) == 0 || len(cond.Roles) == 0 {
-		return false, trace.BadParameter("the system does not support empty roles and users")
-	}
-
-	anyUser, err := parse.NewAnyMatcher(cond.Users)
-	if err != nil {
-		return false, trace.Wrap(err)
-	}
-
-	if anyUser.Match(impersonateUser.GetName()) {
-		return true, nil
-	}
-
-	anyRole, err := parse.NewAnyMatcher(cond.Roles)
-	if err != nil {
-		return false, trace.Wrap(err)
-	}
-
-	for _, impersonateRole := range impersonateRoles {
-		if anyRole.Match(impersonateRole.GetName()) {
-			return true, nil
-		}
-	}
-
-	return false, nil
-}
-
-// RoleMatcher defines an interface for a generic role matcher.
-type RoleMatcher interface {
-	Match(Role, RoleConditionType) (bool, error)
-}
-
-// RoleMatchers defines a list of matchers.
-type RoleMatchers []RoleMatcher
-
-// MatchAll returns true if all matchers in the set match.
-func (m RoleMatchers) MatchAll(role Role, condition RoleConditionType) (bool, error) {
-	for _, matcher := range m {
-		match, err := matcher.Match(role, condition)
-		if err != nil {
-			return false, trace.Wrap(err)
-		}
-		if !match {
-			return false, nil
-		}
-	}
-	return true, nil
-}
-
-// MatchAny returns true if at least one of the matchers in the set matches.
-//
-// If the result is true, returns matcher that matched.
-func (m RoleMatchers) MatchAny(role Role, condition RoleConditionType) (bool, RoleMatcher, error) {
-	for _, matcher := range m {
-		match, err := matcher.Match(role, condition)
-		if err != nil {
-			return false, nil, trace.Wrap(err)
-		}
-		if match {
-			return true, matcher, nil
-		}
-	}
-	return false, nil, nil
-}
-
-// DatabaseLabelsMatcher matches a role against a list of database server labels.
-type DatabaseLabelsMatcher struct {
-	Labels map[string]string
-}
-
-// Match matches database server labels against provided role and condition.
-func (m *DatabaseLabelsMatcher) Match(role Role, condition RoleConditionType) (bool, error) {
-	match, _, err := MatchLabels(role.GetDatabaseLabels(condition), m.Labels)
-	return match, trace.Wrap(err)
-}
-
-// String returns the matcher's string representation.
-func (m *DatabaseLabelsMatcher) String() string {
-	return fmt.Sprintf("DatabaseLabelsMatcher(Labels=%v)", m.Labels)
-}
-
-// DatabaseUserMatcher matches a role against database account name.
-type DatabaseUserMatcher struct {
-	User string
-}
-
-// Match matches database account name against provided role and condition.
-func (m *DatabaseUserMatcher) Match(role Role, condition RoleConditionType) (bool, error) {
-	match, _ := MatchDatabaseUser(role.GetDatabaseUsers(condition), m.User)
-	return match, nil
-}
-
-// String returns the matcher's string representation.
-func (m *DatabaseUserMatcher) String() string {
-	return fmt.Sprintf("DatabaseUserMatcher(User=%v)", m.User)
-}
-
-// DatabaseNameMatcher matches a role against database name.
-type DatabaseNameMatcher struct {
-	Name string
-}
-
-// Match matches database name against provided role and condition.
-func (m *DatabaseNameMatcher) Match(role Role, condition RoleConditionType) (bool, error) {
-	match, _ := MatchDatabaseName(role.GetDatabaseNames(condition), m.Name)
-	return match, nil
-}
-
-// String returns the matcher's string representation.
-func (m *DatabaseNameMatcher) String() string {
-	return fmt.Sprintf("DatabaseNameMatcher(Name=%v)", m.Name)
-}
-
-// CheckAccessToDatabase checks if this role set has access to a particular database.
-//
-// The checker always checks the server namespace, other matchers are supplied
-// by the caller.
-func (set RoleSet) CheckAccessToDatabase(server types.DatabaseServer, mfa AccessMFAParams, matchers ...RoleMatcher) error {
-	log := log.WithField(trace.Component, teleport.ComponentRBAC)
-	if mfa.AlwaysRequired && !mfa.Verified {
-		log.Debugf("Access to database %q denied, cluster requires per-session MFA", server.GetName())
-		return ErrSessionMFARequired
-	}
-	// Check deny rules.
-	for _, role := range set {
-		matchNamespace, _ := MatchNamespace(role.GetNamespaces(Deny), server.GetNamespace())
-		// Deny rules are greedy on purpose. They will always match if
-		// at least one of the matchers returns true.
-		if matchNamespace {
-			match, matcher, err := RoleMatchers(matchers).MatchAny(role, Deny)
-			if err != nil {
-				return trace.Wrap(err)
-			}
-			if match {
-				log.Debugf("Access to database %q denied, deny rule in role %q matched; %s.",
-					server.GetName(), role.GetName(), matcher)
-				return trace.AccessDenied("access to database denied")
-			}
-		}
-	}
-	allowed := false
-	// Check allow rules.
-	for _, role := range set {
-		matchNamespace, _ := MatchNamespace(role.GetNamespaces(Allow), server.GetNamespace())
-		// Allow rules are not greedy. They will match only if all of the
-		// matchers return true.
-		if matchNamespace {
-			match, err := RoleMatchers(matchers).MatchAll(role, Allow)
-			if err != nil {
-				return trace.Wrap(err)
-			}
-			if match {
-				if mfa.Verified {
-					return nil
-				}
-				if role.GetOptions().RequireSessionMFA {
-					log.Debugf("Access to database %q denied, role %q requires per-session MFA", server.GetName(), role.GetName())
-					return ErrSessionMFARequired
-				}
-				// Check all remaining roles, even if we found a match.
-				// RequireSessionMFA should be enforced when at least one role has
-				// it.
-				allowed = true
-				log.Debugf("Access to database %q granted, allow rule in role %q matched.",
-					server.GetName(), role.GetName())
-				continue
-			}
-		}
-	}
-	if allowed {
-		return nil
-	}
-
-	log.Debugf("Access to database %q denied, no allow rule matched.",
-		server.GetName())
-	return trace.AccessDenied("access to database denied")
-}
-
-// CanForwardAgents returns true if role set allows forwarding agents.
-func (set RoleSet) CanForwardAgents() bool {
-	for _, role := range set {
-		if role.GetOptions().ForwardAgent.Value() {
-			return true
-		}
-	}
-	return false
-}
-
-// CanPortForward returns true if a role in the RoleSet allows port forwarding.
-func (set RoleSet) CanPortForward() bool {
-	for _, role := range set {
-		if BoolDefaultTrue(role.GetOptions().PortForwarding) {
-			return true
-		}
-	}
-	return false
-}
-
-// MaybeCanReviewRequests attempts to guess if this RoleSet belongs
-// to a user who should be submitting access reviews.  Because not all rolesets
-// are derived from statically assigned roles, this may return false positives.
-func (set RoleSet) MaybeCanReviewRequests() bool {
-	for _, role := range set {
-		if !role.GetAccessReviewConditions(Allow).IsZero() {
-			// at least one nonzero allow directive exists for
-			// review submission.
-			return true
-		}
-	}
-	return false
-}
-
-// PermitX11Forwarding returns true if this RoleSet allows X11 Forwarding.
-func (set RoleSet) PermitX11Forwarding() bool {
-	for _, role := range set {
-		if role.GetOptions().PermitX11Forwarding.Value() {
-			return true
-		}
-	}
-	return false
-}
-
-// CertificateFormat returns the most permissive certificate format in a
-// RoleSet.
-func (set RoleSet) CertificateFormat() string {
-	var formats []string
-
-	for _, role := range set {
-		// get the certificate format for each individual role. if a role does not
-		// have a certificate format (like implicit roles) skip over it
-		certificateFormat := role.GetOptions().CertificateFormat
-		if certificateFormat == "" {
-			continue
-		}
-
-		formats = append(formats, certificateFormat)
-	}
-
-	// if no formats were found, return standard
-	if len(formats) == 0 {
-		return teleport.CertificateFormatStandard
-	}
-
-	// sort the slice so the most permissive is the first element
-	sort.Slice(formats, func(i, j int) bool {
-		return certificatePriority(formats[i]) < certificatePriority(formats[j])
-	})
-
-	return formats[0]
-}
-
-// EnhancedRecordingSet returns the set of enhanced session recording
-// events to capture for thi role set.
-func (set RoleSet) EnhancedRecordingSet() map[string]bool {
-	m := make(map[string]bool)
-
-	// Loop over all roles and create a set of all options.
-	for _, role := range set {
-		for _, opt := range role.GetOptions().BPF {
-			m[opt] = true
-		}
-	}
-
-	return m
-}
-
-// certificatePriority returns the priority of the certificate format. The
-// most permissive has lowest value.
-func certificatePriority(s string) int {
-	switch s {
-	case teleport.CertificateFormatOldSSH:
-		return 0
-	case teleport.CertificateFormatStandard:
-		return 1
-	default:
-		return 2
-	}
-}
-
-// CheckAgentForward checks if the role can request to forward the SSH agent
-// for this user.
-func (set RoleSet) CheckAgentForward(login string) error {
-	// check if we have permission to login and forward agent. we don't check
-	// for deny rules because if you can't forward an agent if you can't login
-	// in the first place.
-	for _, role := range set {
-		for _, l := range role.GetLogins(Allow) {
-			if role.GetOptions().ForwardAgent.Value() && l == login {
-				return nil
-			}
-		}
-	}
-	return trace.AccessDenied("%v can not forward agent for %v", set, login)
-}
-
-func (set RoleSet) String() string {
-	if len(set) == 0 {
-		return "user without assigned roles"
-	}
-	roleNames := make([]string, len(set))
-	for i, role := range set {
-		roleNames[i] = role.GetName()
-	}
-	return fmt.Sprintf("roles %v", strings.Join(roleNames, ","))
-}
-
-// CheckAccessToRule checks if the RoleSet provides access in the given
-// namespace to the specified resource and verb.
-// silent controls whether the access violations are logged.
-func (set RoleSet) CheckAccessToRule(ctx RuleContext, namespace string, resource string, verb string, silent bool) error {
-	whereParser, err := NewWhereParser(ctx)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-	actionsParser, err := NewActionsParser(ctx)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-	// check deny: a single match on a deny rule prohibits access
-	for _, role := range set {
-		matchNamespace, _ := MatchNamespace(role.GetNamespaces(Deny), ProcessNamespace(namespace))
-		if matchNamespace {
-			matched, err := MakeRuleSet(role.GetRules(Deny)).Match(whereParser, actionsParser, resource, verb)
-			if err != nil {
-				return trace.Wrap(err)
-			}
-			if matched {
-				if !silent {
-					log.WithFields(log.Fields{
-						trace.Component: teleport.ComponentRBAC,
-					}).Infof("Access to %v %v in namespace %v denied to %v: deny rule matched.",
-						verb, resource, namespace, role.GetName())
-				}
-				return trace.AccessDenied("access denied to perform action '%s' on %s", verb, resource)
-			}
-		}
-	}
-
-	// check allow: if rule matches, grant access to resource
-	for _, role := range set {
-		matchNamespace, _ := MatchNamespace(role.GetNamespaces(Allow), ProcessNamespace(namespace))
-		if matchNamespace {
-			match, err := MakeRuleSet(role.GetRules(Allow)).Match(whereParser, actionsParser, resource, verb)
-			if err != nil {
-				return trace.Wrap(err)
-			}
-			if match {
-				return nil
-			}
-		}
-	}
-
-	if !silent {
-		log.WithFields(log.Fields{
-			trace.Component: teleport.ComponentRBAC,
-		}).Infof("Access to %v %v in namespace %v denied to %v: no allow rule matched.",
-			verb, resource, namespace, set)
-	}
-	return trace.AccessDenied("access denied to perform action %q on %q", verb, resource)
-}
-
-// AccessMFAParams contains MFA-related parameters for CheckAccessTo* methods.
-type AccessMFAParams struct {
-	// AlwaysRequired is set when MFA is required for all sessions, regardless
-	// of per-role options.
-	AlwaysRequired bool
-	// Verified is set when MFA has been verified by the caller.
-	Verified bool
 }
 
 // SortedRoles sorts roles by name
